requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
requests==2.31.0
folium==0.15.0
streamlit-folium==0.25.3
scikit-learn==1.3.2
pandas==2.1.3
openpyxl==3.1.2
joblib==1.3.2
python-dotenv==1.0.0
sentinelhub==3.10.0
Pillow==10.1.0
numpy==1.26.4
openai==1.3.0
plotly==5.18.0


